Pulled functions out of Station Component.
Put into a utility function file.
